chore: 파일 정리 및 예제 페이지 작성
