+++ app/example/page.module.css
@@ -1,0 +1,3 @@
+.wow {
+    color:aqua;
+}
