+++ app/example/page.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import styles from "./page.module.css";
+
+export default function Example() {
+  return <div className={styles.wow}>wow</div>;
+}
