+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-app",
+  "name": "chsua-votogether-next-app",
   "version": "0.1.0",
   "private": true,
   "scripts": {
